5.1-RAG_question_answering_langraph/requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
langgraph
jq
pymupdf
redis
flask
markdown
python-dotenv